Move default state in separate file.
